punto dos con bugs

+++ Practicas/Practica4/ruido_gaussiano.m
@@ -3,5 +3,5 @@ function ruido = ruido_gaussiano(img, sigma, mu)
 img = double(img)/255;
 gen = mu + (sigma-mu)*randn(size(img));
 ruido = gen + img;
-ruido = (ruido*255);
+ruido = uint8(ruido*255);
 endfunction
